SampleRasLib: Sample RAS Library
There are certain aspects of RAS which are platform dependent. i.e. How
error records are discovered. This library implements such APIs that
various drivers can use. First version introduces the functionality
to provide an array of all platform discovered RAS error sources. This
array is then used to build required source descriptors as well as used
for error handling. Note that while currently we are scanning for
MEM records based on impid, in future we will push this functionality to
platform FW as it has better handle to understand which records belongs
to which error type.

Signed-off-by: Akshay Behl <cap2k4@rivosinc.com>

Co-authored-by: Dhaval Sharma <dhaval@rivosinc.com>